requirements.txt
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
jsonschema==4.20.0
pandas==2.1.0
openpyxl==3.1.0
pyyaml==6.0.1
scipy==1.11.0
click==8.1.7
colorama==0.4.6
